update handleLogout based on user condition

+++ client/src/components/Header.js
@@ -6,11 +6,15 @@ function Header() {
   const { user, setUser } = useContext(UserContext)
 
   function handleLogout() {
-    fetch('/logout', { method: "DELETE" }).then(r => {
-      if(r.ok) {
-        setUser(null)
-      }
-    })
+    if(user) {
+      fetch('/logout', { method: "DELETE" }).then(r => {
+        if(r.ok) {
+          setUser(null)
+        }
+      })
+    } else {
+      navigate('/loginsignup')
+    }
   }
   
   return (
